\name{graphDataDefaults-methods}
\docType{methods}
\alias{graphDataDefaults-methods}
\alias{graphDataDefaults<--methods}
\alias{graphDataDefaults}
\alias{graphDataDefaults<-}

\title{Get and set default attributes for an Ragraph}

\usage{
graphDataDefaults(self)
graphDataDefaults(self, attr) <- value
}
\arguments{
  \item{self}{A \code{\link{Ragraph-class}} instance}
  \item{attr}{A \code{character} value giving the name of the attribute}
  \item{value}{A \code{character} value as the default value for the specified
    attribute}
}
\description{
  Get/Set default values for attributes associated with a graph.
}

\author{Li Long <li.long@isb-sib.ch>}

\keyword{methods}

