\name{graphvizCapabilities}
\alias{graphvizCapabilities}
\title{
  List capabilities of Graphviz
}
\description{
  List the capabilities of the installed Graphviz.
}
\usage{
graphvizCapabilities()
}
\value{
  A list of 5 character vectors.  Each character vector describes the
  relevant capabilitis of Graphviz.
}
\author{
  Kasper Daniel Hansen \email{khansen@jhsph.edu}
}
\examples{
graphvizCapabilities()
}
