\name{graphvizVersion}
\alias{graphvizVersion}
\title{ A function to determine graphviz library version}
\description{
  This function will query the graphviz libraries that the package was
  built against and report what version of Graphviz is being used.
}
\usage{
graphvizVersion()
}
\value{
  A character string reporting the version}
}
\author{Jeff Gentry}
\examples{
graphvizVersion()
}
\keyword{utilities}
